docs: first prototype to link from events to API reference
