+++ docs/modd.conf
@@ -1,5 +1,5 @@
-scripts/*.py {
-    prep: ./build.sh
+scripts/** {
+    prep: python3 build.py
 }
 
 {
