+++ docs/src/assets/style.scss
@@ -7,6 +7,9 @@ $family-sans-serif: BlinkMacSystemFont, -apple-system, "Segoe UI", "Roboto", "Ox
 
 $panel-heading-size: 1em;
 $panel-heading-weight: 600;
+$menu-list-link-padding: .3em .75em;
+$menu-label-spacing: .7em;
+$menu-nested-list-margin: .3em .75em;
 
 /*!*
 bulma.io v0.8.0 | MIT License | github.com/jgthms/bulma */
