+++ docs/src/content/addons-scripting.md
@@ -21,7 +21,7 @@ an arbitrary response instead:
 
 {{< example src="examples/addons/http-reply-from-proxy.py" lang="py" >}}
 
-All events around the HTTP protocol [can be found here]({{< relref "addons-events#http-events">}}).
+All events around the HTTP protocol [can be found here]({{< relref "addons-api#http-events">}}).
 
 For HTTP-related objects, please look at the [http][] module, or the
 [Request][], and [Response][] classes for other attributes that you can use when
@@ -29,7 +29,7 @@ scripting.
 
 # Scripting WebSocket
 
-The WebSocket protocol initially looks like a regular HTTP request, before the client and server agree to upgrade the connection to WebSocket. All scripting events for initial HTTP handshake, and also the dedicated WebSocket events [can be found here]({{< relref "addons-events#websocket-events">}}).
+The WebSocket protocol initially looks like a regular HTTP request, before the client and server agree to upgrade the connection to WebSocket. All scripting events for initial HTTP handshake, and also the dedicated WebSocket events [can be found here]({{< relref "addons-api#websocket-events">}}).
 
 {{< example src="examples/addons/websocket-simple.py" lang="py" >}}
 
@@ -41,7 +41,7 @@ all attributes that you can use when scripting.
 
 # Scripting TCP
 
-All events around the TCP protocol [can be found here]({{< relref "addons-events#tcp-events">}}).
+All events around the TCP protocol [can be found here]({{< relref "addons-api#tcp-events">}}).
 
 {{< example src="examples/addons/tcp-simple.py" lang="py" >}}
