--- docs/src/content/api/_index.md
@@ -1,10 +1,0 @@
-
----
-title: "API Reference"
-layout: single
-menu:
-    addons:
-        weight: 5
----
-
-# API Reference
