+++ docs/src/content/api/mitmproxy.websocket.md
@@ -5,7 +5,7 @@ url: "api/mitmproxy/websocket.html"
 
 menu:
     addons:
-        parent: 'API Reference'
+        parent: 'API'
 ---
 
 {{< readfile file="/generated/api/mitmproxy/websocket.html" >}}
